Rename and move user remote nested object class
